Refactor all tests to use Testify
Ginkgo/Gomega was a lot to handle every time we revisit this project.
It's the only project that uses this test framework (which was an
experiment in the first place...) and using the DSL/cascading functions
isn't the easiest to bring someone up quickly for a small change.
Given how infrequently we need to visit this codebase, I think we should
follow KISS here and just use a small helper.

Testify is a simple assertions/suite wrapper. It just takes a little bit
of the edge off native Go testing, while still being very simple.