Fix inventory target submatch
Because we used `strings.Contains(x, y)` for determining the correct
inventory to use, a submatch would yield the wrong inventory sometimes.

This was an issue if we were looking for `10.0.0.1`, but `10.0.0.10` was
in an inventory that was checked before the correct inventory. The
former is a substring of the latter.

This change makes the check against the entire line.