Fix build (#40)
The native go build system is broken for this repo, thus the dependencies are no longer managed
correctly with `go module`.  Though the go version has been changed to go 1.19, dependencies are
not updated accordingly.

In this change:

- pin bazel to 5.4.1 with `.bazelversion`
- update `go.mod` to use `go 1.19` to match `WORKSPACE`
- run `go mod tidy` to clean up `go.mod` and `go.sum` for new version
- update `github.com/spf13/viper` to 1.16.0 to resolve compatibility issues
- update `bazel_gazelle` to the latest version (v0.31.1) that is compatible with `rules_go` 0.35.0
- update build files with `bazel run //:gazelle` and remove redundant test targets
- add `//:update-deps` target to manage `deps.bzl`
- regenerate `deps.bzl` with `bazel run //:update-deps`

The change should be a no-op, as versions of direct dependencies that provide core functionalities
are remain unchanged.