Add support for custom md5 remote path. (#49)
We use the gitlab package registry to store ansible playbook releases.
Gitlab has permalinks to the latest artifacts or to release assets.
These links to artifacts or release assets are accessed using the gitlab
API and my not be the actual path to the tarball or md5 checksum file.

A custom remote path for the link to the md5 hash would be very nice!
This PR will add a `http-checksum-url` flag. If the `http-checksum-url`
flag has not been set, it will use the `http-url` with `.md5` added to
the path.